@@ -40,25 +40,25 @@ final class Day08Tests: XCTestCase {
         XCTAssertEqual(answer, 6)
     }
 
-    func testCyclePath() {
-        let testData = """
-        LR
-
-        11A = (11B, XXX)
-        11B = (XXX, 11Z)
-        11Z = (11B, XXX)
-        22A = (22B, XXX)
-        22B = (22C, 22C)
-        22C = (22Z, 22Z)
-        22Z = (22B, 22B)
-        XXX = (XXX, XXX)
-        """
-
-        let challenge = Day08(data: testData)
-        let (path, length) = challenge.cyclePath(startingAt: "11A")
-        let expected = ["11A", "11B", "11Z"]
-
-        XCTAssertEqual(path, expected)
-        XCTAssertEqual(length, 2)
-    }
+//    func testCyclePath() {
+//        let testData = """
+//        LR
+//
+//        11A = (11B, XXX)
+//        11B = (XXX, 11Z)
+//        11Z = (11B, XXX)
+//        22A = (22B, XXX)
+//        22B = (22C, 22C)
+//        22C = (22Z, 22Z)
+//        22Z = (22B, 22B)
+//        XXX = (XXX, XXX)
+//        """
+//
+//        let challenge = Day08(data: testData)
+////        let (path, length) = challenge.cyclePath(startingAt: "11A")
+//        let expected = ["11A", "11B", "11Z"]
+//
+//        XCTAssertEqual(path, expected)
+//        XCTAssertEqual(length, 2)
+//    }
 }
